backend-python/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "firebase-admin>=7.4.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
    "pystray>=0.19.5",
    "Pillow>=10.4.0",
]
